+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Hint to Page Form.
+- Validate if URL begins with '/'.
 
 ### Fixed